fix: removed test feature for upload

+++ donzoby.com/app/Http/Controllers/PostSyncController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
-use App\Models\Test_post;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
@@ -16,16 +15,16 @@ class PostSyncController extends Controller
             // Log::info('Received data::' . json_encode($request->all()));
             if ($request->has('is_new_post') || $request->has('just_syncing')) {
                 // if post already exist for just_syncing action, just update
-                $post_exists = Test_post::find($request->id);
+                $post_exists = Post::find($request->id);
                 if ($post_exists) {
                     Log::info('^^^^^^^^^^^^^Just syncing and post exists^^^^^^^^^^');
                     $post_exists->update($request->toArray());
                 } else {
-                    $last_post_id = Test_post::latest('id')->first()->id;
+                    $last_post_id = Post::latest('id')->first()->id;
 
                     // add dummy (unpublished) posts till $id-1 while last post id + 1 < $request->id
                     while ($last_post_id + 1 < $request->id) {
-                        $dummy_post = Test_post::create([
+                        $dummy_post = Post::create([
                             "topic" => "dummy",
                             "content" => "dummy",
                             "status" => "unpublished",
@@ -46,12 +45,12 @@ class PostSyncController extends Controller
                             'message' => 'Post out of sync. Next post id is ' . ($last_post_id + 1) . ' and received id is ' . $request->id . '.',
                         ], 422);
                     }
-                    $post = Test_post::create($request->toArray() + ['author_id' => 1]);
+                    $post = Post::create($request->toArray() + ['author_id' => 1]);
                 }
             } else { // it is post edit
                 // update the post
                 Log::info('::::::::::::::::::::::::RECEIVED POST ID:::::::::::::::::::::::::' . $request->id);
-                $post = Test_post::where('id', $request->id)->first();
+                $post = Post::where('id', $request->id)->first();
                 if (!$post) {
                     return response()->json([
                         'status' => 'error',
@@ -70,7 +69,7 @@ class PostSyncController extends Controller
             // ****Down the line, I will need to find a fix for that but for now, the below works
             foreach ($request->file() as $image) {
                 Log::info('+++++++++++++++++++IMAGE FILE FOUND+++++++++++++++++++');
-                $image->move("images/courses/xphp/{$post->subject->slug}", $image->getClientOriginalName());
+                $image->move("images/courses/{$post->subject->slug}", $image->getClientOriginalName());
                 // add to post images
                 $post->post_images()->create(['link' => "images/courses/{$post->subject->slug}/{$image->getClientOriginalName()}"]);
             }
@@ -79,10 +78,10 @@ class PostSyncController extends Controller
                 var_dump($request->removed_images);
                 foreach ($request->removed_images as $image) {
                     Log::info('------Removed image::' . json_encode($request->removed_images));
-                    /* if (file_exists($image)) {
+                    if (file_exists($image)) {
                         Log::info('found image file and about to delete:::' . $image);
                         unlink($image);
-                    } */
+                    }
                 }
             }
             return [
